pyproject.toml
[project]
name = "yt"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pytubefix>=10.3.6",
]

[dependency-groups]
dev = [
    "basedpyright>=1.36.0",
    "prek>=0.2.21",
    "ruff>=0.14.8",
]

[tool.basedpyright]
reportExplicitAny = false

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D"]
